Test: split monolithic MultibindingTest to MapMultibindingTest & SetMultibindingTest
