+++ projects/core/koin-core/src/commonTest/kotlin/org/koin/core/MapMultibindingTest.kt
@@ -3,7 +3,6 @@ package org.koin.core
 import co.touchlab.stately.concurrency.AtomicInt
 import org.koin.Simple
 import org.koin.core.module.MapMultibindingKeyTypeException
-import org.koin.core.parameter.parameterSetOf
 import org.koin.core.parameter.parametersOf
 import org.koin.core.qualifier._q
 import org.koin.core.qualifier.named
@@ -16,7 +15,7 @@ import kotlin.test.assertFailsWith
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
 
-class MultibindingTest {
+class MapMultibindingTest {
     private val keyOfComponent1 = "component1"
     private val keyOfComponent2 = "component2"
     private val component1 = Simple.Component1()
@@ -62,7 +61,7 @@ class MultibindingTest {
     }
 
     @Test
-    fun `inject some elements into map multibinding in root scope`() {
+    fun `inject elements into map multibinding in root scope`() {
         val app = koinApplication {
             modules(
                 module {
@@ -88,7 +87,7 @@ class MultibindingTest {
     }
 
     @Test
-    fun `inject some elements into map multibinding in none root scope`() {
+    fun `inject elements into map multibinding in none root scope`() {
         val app = koinApplication {
             modules(
                 module {
@@ -285,242 +284,24 @@ class MultibindingTest {
     }
 
     @Test
-    fun `declare set multibinding in root scope`() {
+    fun `declare map multibinding elements in separated modules`() {
         val app = koinApplication {
             modules(
                 module {
-                    declareSetMultibinding<Simple.ComponentInterface1>()
-                },
-            )
-        }
-
-        val koin = app.koin
-        val set: Set<Simple.ComponentInterface1> = koin.getSetMultibinding()
-        val set2: Set<Simple.ComponentInterface1> = koin.getSetMultibinding()
-        assertEquals(set, set2)
-        assertTrue { set.isEmpty() }
-    }
-
-    @Test
-    fun `declare set multibinding in none root scope`() {
-        val app = koinApplication {
-            modules(
-                module {
-                    scope(scopeKey) {
-                        declareSetMultibinding<Simple.ComponentInterface1>()
-                    }
-                },
-            )
-        }
-
-        val koin = app.koin
-        val myScope = koin.createScope(scopeId, scopeKey)
-        val set: Set<Simple.ComponentInterface1> = myScope.getSetMultibinding()
-        val set2: Set<Simple.ComponentInterface1> = myScope.getSetMultibinding()
-        assertEquals(set, set2)
-        assertTrue { set.isEmpty() }
-    }
-
-    @Test
-    fun `inject some elements into set multibinding in root scope`() {
-        val app = koinApplication {
-            modules(
-                module {
-                    declareSetMultibinding<Simple.ComponentInterface1> {
-                        intoSet { component1 }
-                        intoSet { component2 }
-                    }
-                },
-            )
-        }
-
-        val koin = app.koin
-        val set: Set<Simple.ComponentInterface1> = koin.getSetMultibinding()
-        val set2: Set<Simple.ComponentInterface1> = koin.getSetMultibinding()
-        assertEquals(set, set2)
-
-        assertEquals(2, set.size)
-        assertTrue { set.containsAll(listOf(component1, component2)) }
-    }
-
-    @Test
-    fun `inject some elements into set multibinding in none root scope`() {
-        val app = koinApplication {
-            modules(
-                module {
-                    scope(scopeKey) {
-                        declareSetMultibinding<Simple.ComponentInterface1> {
-                            intoSet { component1 }
-                            intoSet { component2 }
-                        }
-                    }
-                },
-            )
-        }
-
-        val koin = app.koin
-        val myScope = koin.createScope(scopeId, scopeKey)
-        val set: Set<Simple.ComponentInterface1> = myScope.getSetMultibinding()
-        val set2: Set<Simple.ComponentInterface1> = myScope.getSetMultibinding()
-        assertEquals(set, set2)
-
-        assertEquals(2, set.size)
-        assertTrue { set.containsAll(listOf(component1, component2)) }
-    }
-
-    @Test
-    fun `set multibinding contains all elements that define in linked scope`() {
-        val rootComponent1 = Simple.Component1()
-        val rootComponent2 = Simple.Component2()
-        val scopeComponent1 = Simple.Component1()
-        val scopeComponent2 = Simple.Component2()
-        val app = koinApplication {
-            modules(
-                module {
-                    declareSetMultibinding<Simple.ComponentInterface1> {
-                        intoSet { rootComponent1 }
-                        intoSet { rootComponent2 }
-                    }
-                    scope(scopeKey) {
-                        declareSetMultibinding<Simple.ComponentInterface1> {
-                            intoSet { scopeComponent1 }
-                            intoSet { scopeComponent2 }
-                        }
-                    }
-                },
-            )
-        }
-
-        val koin = app.koin
-        val myScope = koin.createScope(scopeId, scopeKey)
-        val rootSet: Set<Simple.ComponentInterface1> = koin.getSetMultibinding()
-        val scopeSet: Set<Simple.ComponentInterface1> = myScope.getSetMultibinding()
-
-        assertEquals(2, rootSet.size)
-        assertTrue { rootSet.containsAll(listOf(rootComponent1, rootComponent2)) }
-        assertEquals(4, scopeSet.size)
-        assertTrue {
-            scopeSet.containsAll(
-                listOf(
-                    rootComponent1,
-                    rootComponent2,
-                    scopeComponent1,
-                    scopeComponent2,
-                )
-            )
-        }
-        // in elements definition order
-        assertEquals(
-            listOf(
-                rootComponent1,
-                rootComponent2,
-                scopeComponent1,
-                scopeComponent2,
-            ),
-            scopeSet.toList()
-        )
-    }
-
-    @Test
-    fun `override set multibinding elements`() {
-        data class SetElement(
-            val intValue: Int = 0
-        ) {
-            var name = ""
-        }
-
-        val app = koinApplication {
-            modules(
-                module {
-                    declareSetMultibinding<SetElement> {
-                        intoSet { SetElement().apply { name = "element1" } }
-                        intoSet { SetElement().apply { name = "element2" } }
-                    }
-                },
-            )
-        }
-
-        val koin = app.koin
-        val rootSet: Set<SetElement> = koin.getSetMultibinding()
-        assertEquals(1, rootSet.size)
-        assertEquals("element2", (rootSet.first() as SetElement).name)
-        koin.loadModules(
-            listOf(
-                module {
-                    declareSetMultibinding<SetElement> {
-                        intoSet { SetElement().apply { name = "element1" } }
-                    }
-                }
-            )
-        )
-        assertEquals(1, rootSet.size)
-        assertEquals("element1", (rootSet.first() as SetElement).name)
-    }
-
-    @Test
-    fun `override set multibinding elements that define in linked scope`() {
-        data class SetElement(
-            val intValue: Int = 0
-        ) {
-            var name = ""
-        }
-
-        val app = koinApplication {
-            modules(
-                module {
-                    declareSetMultibinding<SetElement> {
-                        intoSet { SetElement().apply { name = "root" } }
-                    }
-                    scope(scopeKey) {
-                        declareSetMultibinding<SetElement> {
-                            intoSet { SetElement().apply { name = "scoped" } }
-                        }
-                    }
-                },
-            )
-        }
-
-        val koin = app.koin
-        val myScope = koin.createScope(scopeId, scopeKey)
-        val rootSet: Set<SetElement> = koin.getSetMultibinding()
-        val scopeSet: Set<SetElement> = myScope.getSetMultibinding()
-        assertEquals(1, rootSet.size)
-        assertEquals(1, scopeSet.size)
-        assertEquals("root", (rootSet.first() as SetElement).name)
-        assertEquals("scoped", (scopeSet.first() as SetElement).name)
-    }
-
-    @Test
-    fun `declare multibinding elements in separated modules`() {
-        val app = koinApplication {
-            modules(
-                module {
-                    declareSetMultibinding<Simple.ComponentInterface1> {
-                        intoSet { component1 }
-                    }
                     declareMapMultibinding<String, Simple.ComponentInterface1> {
                         intoMap(keyOfComponent1) { component1 }
                     }
                     scope(scopeKey) {
-                        declareSetMultibinding<Simple.ComponentInterface1> {
-                            intoSet { component1 }
-                        }
                         declareMapMultibinding<String, Simple.ComponentInterface1> {
                             intoMap(keyOfComponent1) { component1 }
                         }
                     }
                 },
                 module {
-                    declareSetMultibinding<Simple.ComponentInterface1> {
-                        intoSet { component2 }
-                    }
                     declareMapMultibinding<String, Simple.ComponentInterface1> {
                         intoMap(keyOfComponent2) { component2 }
                     }
                     scope(scopeKey) {
-                        declareSetMultibinding<Simple.ComponentInterface1> {
-                            intoSet { component2 }
-                        }
                         declareMapMultibinding<String, Simple.ComponentInterface1> {
                             intoMap(keyOfComponent2) { component2 }
                         }
@@ -531,15 +312,9 @@ class MultibindingTest {
 
         val koin = app.koin
         val myScope = koin.createScope(scopeId, scopeKey)
-        val rootSet: Set<Simple.ComponentInterface1> = koin.getSetMultibinding()
-        val scopeSet: Set<Simple.ComponentInterface1> = myScope.getSetMultibinding()
         val rootMap: Map<String, Simple.ComponentInterface1> = koin.getMapMultibinding()
         val scopeMap: Map<String, Simple.ComponentInterface1> = myScope.getMapMultibinding()
 
-        for (set in listOf(rootSet, scopeSet)) {
-            assertEquals(2, set.size)
-            assertTrue { set.containsAll(listOf(component1, component2)) }
-        }
         for (map in listOf(rootMap, scopeMap)) {
             assertEquals(2, map.size)
             assertTrue { map.values.containsAll(listOf(component1, component2)) }
@@ -547,16 +322,10 @@ class MultibindingTest {
     }
 
     @Test
-    fun `declare multibinding elements through different MultibindingElementDefinitions`() {
+    fun `declare map multibinding elements through different MultibindingElementDefinitions`() {
         val app = koinApplication {
             modules(
                 module {
-                    declareSetMultibinding<Simple.ComponentInterface1> {
-                        intoSet { component1 }
-                    }
-                    declareSetMultibinding<Simple.ComponentInterface1> {
-                        intoSet { component2 }
-                    }
                     declareMapMultibinding<String, Simple.ComponentInterface1> {
                         intoMap(keyOfComponent1) { component1 }
                     }
@@ -564,12 +333,6 @@ class MultibindingTest {
                         intoMap(keyOfComponent2) { component2 }
                     }
                     scope(scopeKey) {
-                        declareSetMultibinding<Simple.ComponentInterface1> {
-                            intoSet { component1 }
-                        }
-                        declareSetMultibinding<Simple.ComponentInterface1> {
-                            intoSet { component2 }
-                        }
                         declareMapMultibinding<String, Simple.ComponentInterface1> {
                             intoMap(keyOfComponent1) { component1 }
                         }
@@ -583,15 +346,9 @@ class MultibindingTest {
 
         val koin = app.koin
         val myScope = koin.createScope(scopeId, scopeKey)
-        val rootSet: Set<Simple.ComponentInterface1> = koin.getSetMultibinding()
-        val scopeSet: Set<Simple.ComponentInterface1> = myScope.getSetMultibinding()
         val rootMap: Map<String, Simple.ComponentInterface1> = koin.getMapMultibinding()
         val scopeMap: Map<String, Simple.ComponentInterface1> = myScope.getMapMultibinding()
 
-        for (set in listOf(rootSet, scopeSet)) {
-            assertEquals(2, set.size)
-            assertTrue { set.containsAll(listOf(component1, component2)) }
-        }
         for (map in listOf(rootMap, scopeMap)) {
             assertEquals(2, map.size)
             assertTrue { map.values.containsAll(listOf(component1, component2)) }
@@ -599,23 +356,15 @@ class MultibindingTest {
     }
 
     @Test
-    fun `declare multibinding elements by MultibindingElementDefinition reference`() {
+    fun `declare map multibinding elements by MultibindingElementDefinition reference`() {
         val app = koinApplication {
             modules(
                 module {
-                    val setElementDefinition =
-                        declareSetMultibinding<Simple.ComponentInterface1>()
-                    setElementDefinition.intoSet { component1 }
-                    setElementDefinition.intoSet { component2 }
                     val mapElementDefinition =
                         declareMapMultibinding<String, Simple.ComponentInterface1>()
                     mapElementDefinition.intoMap(keyOfComponent1) { component1 }
                     mapElementDefinition.intoMap(keyOfComponent2) { component2 }
                     scope(scopeKey) {
-                        val setElementDefinition =
-                            declareSetMultibinding<Simple.ComponentInterface1>()
-                        setElementDefinition.intoSet { component1 }
-                        setElementDefinition.intoSet { component2 }
                         val mapElementDefinition =
                             declareMapMultibinding<String, Simple.ComponentInterface1>()
                         mapElementDefinition.intoMap(keyOfComponent1) { component1 }
@@ -627,15 +376,9 @@ class MultibindingTest {
 
         val koin = app.koin
         val myScope = koin.createScope(scopeId, scopeKey)
-        val rootSet: Set<Simple.ComponentInterface1> = koin.getSetMultibinding()
-        val scopeSet: Set<Simple.ComponentInterface1> = myScope.getSetMultibinding()
         val rootMap: Map<String, Simple.ComponentInterface1> = koin.getMapMultibinding()
         val scopeMap: Map<String, Simple.ComponentInterface1> = myScope.getMapMultibinding()
 
-        for (set in listOf(rootSet, scopeSet)) {
-            assertEquals(2, set.size)
-            assertTrue { set.containsAll(listOf(component1, component2)) }
-        }
         for (map in listOf(rootMap, scopeMap)) {
             assertEquals(2, map.size)
             assertTrue { map.values.containsAll(listOf(component1, component2)) }
@@ -643,63 +386,39 @@ class MultibindingTest {
     }
 
     @Test
-    fun `declare multibinding with specific qualifier`() {
+    fun `declare map multibinding with specific qualifier`() {
         val app = koinApplication {
             modules(
                 module {
-                    declareSetMultibinding<Simple.ComponentInterface1>(_q<Set<*>>()) {
-                        intoSet { component1 }
-                        intoSet { component2 }
-                    }
-                    declareMapMultibinding<String, Simple.ComponentInterface1>(_q("map")) {
+                    declareMapMultibinding<String, Simple.ComponentInterface1>(_q("map1")) {
                         intoMap(keyOfComponent1) { component1 }
                         intoMap(keyOfComponent2) { component2 }
                     }
-                    scope(scopeKey) {
-                        declareSetMultibinding<Simple.ComponentInterface1>(_q<Set<*>>()) {
-                            intoSet { component1 }
-                            intoSet { component2 }
-                        }
-                        declareMapMultibinding<String, Simple.ComponentInterface1>(_q("map1")) {
-                            intoMap(keyOfComponent1) { component1 }
-                            intoMap(keyOfComponent2) { component2 }
-                        }
+                    declareMapMultibinding<String, Simple.ComponentInterface1>(_q("map2")) {
+                        intoMap(keyOfComponent1) { component1 }
+                        intoMap(keyOfComponent2) { component2 }
                     }
                 },
             )
         }
 
         val koin = app.koin
-        val myScope = koin.createScope(scopeId, scopeKey)
-        val rootSet: Set<Simple.ComponentInterface1> =
-            koin.getSetMultibinding(_q<Set<*>>())
-        val scopeSet: Set<Simple.ComponentInterface1> =
-            myScope.getSetMultibinding(_q<Set<*>>())
-        val rootMap: Map<String, Simple.ComponentInterface1> =
-            koin.getMapMultibinding(_q("map"))
-        val scopeMap: Map<String, Simple.ComponentInterface1> =
-            myScope.getMapMultibinding(_q("map1"))
-
-        for (set in listOf(rootSet, scopeSet)) {
-            assertEquals(2, set.size)
-            assertTrue { set.containsAll(listOf(component1, component2)) }
-        }
-        for (map in listOf(rootMap, scopeMap)) {
+        val map1: Map<String, Simple.ComponentInterface1> =
+            koin.getMapMultibinding(_q("map1"))
+        val map2: Map<String, Simple.ComponentInterface1> =
+            koin.getMapMultibinding(_q("map2"))
+        for (map in listOf(map1, map2)) {
             assertEquals(2, map.size)
             assertTrue { map.values.containsAll(listOf(component1, component2)) }
         }
     }
 
     @Test
-    fun `create multibinding at start`() {
+    fun `create map multibinding at start`() {
         val accumulator = AtomicInt(0)
         val app = koinApplication {
             modules(
                 module {
-                    declareSetMultibinding<Int>(createdAtStart = true) {
-                        intoSet { accumulator.incrementAndGet() }
-                        intoSet { accumulator.incrementAndGet() }
-                    }
                     declareMapMultibinding<String, Int>(createdAtStart = true) {
                         intoMap("one") { accumulator.incrementAndGet() }
                         intoMap("two") { accumulator.incrementAndGet() }
@@ -708,21 +427,15 @@ class MultibindingTest {
             )
         }
 
-        val koin = app.koin
-        koin.getSetMultibinding<Int>()
-        koin.getMapMultibinding<String, Int>()
-        assertEquals(4, accumulator.get())
+        app.koin.getMapMultibinding<String, Int>()
+        assertEquals(2, accumulator.get())
     }
 
     @Test
-    fun `create multibinding elements using parameters`() {
+    fun `create map multibinding elements using parameters`() {
         val app = koinApplication {
             modules(
                 module {
-                    declareSetMultibinding<String> {
-                        intoSet { it.get<String>() + "1" }
-                        intoSet { it.get<String>() + "2" }
-                    }
                     declareMapMultibinding<String, String> {
                         intoMap("one") { it.get<String>() + "1" }
                         intoMap("two") { it.get<String>() + "2" }
@@ -731,13 +444,8 @@ class MultibindingTest {
             )
         }
 
-        val koin = app.koin
         assertTrue {
-            koin.getSetMultibinding<String> { parametersOf("set") }
-                .containsAll(listOf("set1", "set2"))
-        }
-        assertTrue {
-            koin.getMapMultibinding<String, String> { parametersOf("map") }
+            app.koin.getMapMultibinding<String, String> { parametersOf("map") }
                 .values.containsAll(listOf("map1", "map2"))
         }
     }
